fix signup and login widths, add more folder structure
