+++ lib/core/services/vote.dart
@@ -2,7 +2,6 @@ import 'dart:convert';
 import 'dart:async' show Future;
 import 'package:flutter/services.dart' show rootBundle;
 import 'package:http/http.dart' as http;
-import 'dart:convert';
 import 'dart:async';
 import 'dart:math';
