+++ lib/ui/screens/bills/pdf_viewer.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:pdf_viewer_plugin/pdf_viewer_plugin.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:http/http.dart' as http;
-import 'package:voting_app/styles.dart';
+import 'package:voting_app/ui/styles.dart';
 
 /// This is the page that displays the PDFs
