+++ lib/ui/screens/login.dart
@@ -1,18 +1,20 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:voting_app/widgets/custom_widgets.dart';
-import 'package:voting_app/styles.dart';
-import 'package:voting_app/screens/login.dart';
+import 'package:voting_app/ui/widgets/custom_widgets.dart';
+import 'package:voting_app/ui/styles.dart';
+import 'package:voting_app/ui/screens/login/signup.dart';
+import 'package:provider/provider.dart';
 
+class ProfilePage extends StatefulWidget {
 
-class SignUpPage extends StatefulWidget {
-  
   @override
-  _SignUpPageState createState() => _SignUpPageState();
+  _ProfilePageState createState() => _ProfilePageState();
 }
 
-class _SignUpPageState extends State<SignUpPage> {
+class _ProfilePageState extends State<ProfilePage> {
   //  Need a method to show we are logged in. json file?
+  bool loggedIn = false;
+  bool login = false;
 
   @override
   Widget build(BuildContext context) {
@@ -20,25 +22,26 @@ class _SignUpPageState extends State<SignUpPage> {
     return Scaffold(
       resizeToAvoidBottomInset: true,
       backgroundColor: appColors.background,
-      body: SignUp()
+      body: (Center(
+        child: Container(
+          width: appSizes.mediumWidth,
+          child: LogIn(),
+        ),
+      )),
     );
   }
 }
 
-class SignUp extends StatefulWidget {
+class LogIn extends StatefulWidget {
   @override
-  _SignUpWidgetState createState() => _SignUpWidgetState();
+  _LogInWidgetState createState() => _LogInWidgetState();
 }
 
-class _SignUpWidgetState extends State<SignUp> {
+class _LogInWidgetState extends State<LogIn> {
 
   final _formKey = GlobalKey<FormState>();
   final Map<String, dynamic> formData = {
-    'first_name': null,
-    'other_given_names': null,
-    'family_name': null,
     'email': null,
-    'post_code': null,
     'password': null
   };
 
@@ -53,27 +56,9 @@ class _SignUpWidgetState extends State<SignUp> {
             Padding(
               padding: EdgeInsets.only(bottom: 20.0),
               child: Text(
-                "Sign up",
+                "Log In",
                 style: appTextStyles.heading,
               ),
-            ),
-            CustomFormField(
-              helpText: "First Name",
-              submitAction: (String value) {
-                formData['first_name'] = value;
-              },
-            ),
-            CustomFormField(
-              helpText: "Other Given Names",
-              submitAction: (String value) {
-                formData['other_given_names'] = value;
-              },
-            ),
-            CustomFormField(
-              helpText: "Family Name",
-              submitAction: (String value) {
-                formData['family_name'] = value;
-              },
             ),
             CustomFormField(
                 helpText: "Email",
@@ -88,12 +73,6 @@ class _SignUpWidgetState extends State<SignUp> {
                   }
                 }),
             CustomFormField(
-              helpText: "Post Code",
-              submitAction: (String value) {
-                formData['post_code'] = value;
-              },
-            ),
-            CustomFormField(
                 helpText: "Password",
                 submitAction: (String value) {
                   formData['password'] = value;
@@ -103,7 +82,11 @@ class _SignUpWidgetState extends State<SignUp> {
                     return 'Enter a password';
                   }
                 }),
-            _buildSubmitButton()
+            _buildSubmitButton(),
+            new Divider(
+              color: appColors.text,
+            ),
+            _buildSignUpButton()
           ],
         ),
       ),
@@ -115,14 +98,40 @@ class _SignUpWidgetState extends State<SignUp> {
       onPressed: () {
         _submitForm();
       },
-      child: Text('Sign Up'),
+      child: Text('Login'),
+    );
+  }
+
+  Widget _buildSignUpButton() {
+    return RaisedButton(
+      onPressed: () {
+        Navigator.push(
+          context,
+          MaterialPageRoute(
+              builder: (context) => SignUpPage()),
+        );
+      },
+      child: Text('Sign Up Now'),
     );
   }
 
   void _submitForm() {
+    print("Fix errors");
     if (_formKey.currentState.validate()) {
       _formKey.currentState.save(); //onSaved is called!
       print(formData);
     }
+  }
+}
+
+class ProfileWidget extends StatefulWidget {
+  @override
+  _ProfileWidgetState createState() => _ProfileWidgetState();
+}
+
+class _ProfileWidgetState extends State<ProfileWidget> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
   }
 }
