+++ lib/ui/widgets/custom_widgets.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:voting_app/styles.dart';
+import 'package:voting_app/ui/styles.dart';
 import 'package:pie_chart/pie_chart.dart';
 import 'dart:async';
