+++ pubspec.yaml
@@ -29,6 +29,7 @@ dependencies:
   path_provider: ^1.6.5
   http: ^0.12.0+4
   pie_chart: ^3.1.1
+  provider: ^4.0.4
 
 dev_dependencies:
   flutter_test:
